test(document-check-function): fixes tests (BOAS-536)
