+++ apps/api/package.json
@@ -38,13 +38,13 @@
     "compression": "1.7.4",
     "cors": "2.8.5",
     "date-fns": "2.28.0",
-    "dotenv": "16.0.1",
+    "dotenv": "16.0.3",
     "express": "4.18.1",
     "express-pino-logger": "^7.0.0",
     "express-validator": "6.14.2",
     "got": "^12.5.2",
     "helmet": "5.1.0",
-    "joi": "17.6.0",
+    "joi": "^17.7.0",
     "lodash-es": "^4.17.21",
     "morgan": "1.10.0",
     "multer": "1.4.4",
