+++ apps/document-storage/package.json
@@ -31,7 +31,7 @@
         "swagger-ui-express": "^4.3.0",
         "compression": "1.7.4",
         "helmet": "5.0.2",
-        "joi": "17.6.0"
+        "joi": "^17.7.0"
     },
     "devDependencies": {
         "ava": "^4.2.0",
